fix parsing
note: there is no way to differentiate a seqset with a single value from
a number. Also we want to parse u64 number, not u32 like currently.
So in the end, I think we will have to parse selectively: when the key
is X, then the value has format Y, etc.